Fixed loading of 'number of internal divisions' segment property from NeuroML files.
Previously, the number of internal divisions of a segment was just silently
ignored when loading a pure NeuroML morphology file. As a side effect, this
property was _removed_ from the morphology if the model was modified and the 
NeuroML file overwritten. Note that, at present, the older (deprecated) form
of the 'property' element (with 'tag' and 'value' as child elements) is not
recognised.